Move ElementIcon into dim-ui
