+++ src/app/dim-ui/ElementIcon.m.scss
@@ -7,9 +7,4 @@
     width: calc(var(--item-size) / 5);
     height: calc(var(--item-size) / 5);
   }
-
-  &.lightBackground {
-    filter: brightness(75%) saturate(200%);
-    background-color: transparent !important;
-  }
 }
