+++ src/app/dim-ui/ElementIcon.m.scss.d.ts
@@ -2,7 +2,6 @@
 // Please do not change this file!
 interface CssExports {
   'element': string;
-  'lightBackground': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
